+++ CHANGELOG.md
@@ -1,8 +1,16 @@
 # CHANGELOG
 
+## v1.1.0 (Unreleased)
+
+-   Features
+    -   `limit` argument: Pass an integer limit of results to return in order to short-circuit and return early from long-running jobs.
+
 ## v1.0.0 (January 11 2021)
 
 > This version adds map hinting in order to let a user specify partial matches in the `find_motifs` call.
+
+-   Features
+    -   `hints` argument: Pass a list of partial starting maps in order to condition the search results.
 
 ## v0.1.0